+++ Changes.md
@@ -1,3 +1,4 @@
+* #1106 - Ensure no duplicate rows in most recent scores scoreboard display
 * #1105 - Rework dynamic display to properly work with stand alone scoreboard pages
 * Better handling of unknown displays