+++ Changes.md
@@ -1,3 +1,5 @@
+Release 18.1.0
+==============
 * #1108 - allow one to manipulate slideshow images from the web
 * Copy slideshow and sponsor logs during migration from an older version of the software
 * #1107 - allow one to manipulate sponsor logos from the web